requirements.txt
mcp
google-auth
httpx
google-adk
pyyaml
